`Recipe.transform` maintains input order for in-memory data
Since Scikit-Learn passes in ``X`` and ``y`` as separate in-memory
objects, we'll want to ``transform`` to maintain the order of ``X`` when
using in-memory data. Some steps may result in reordering (backend
implementation dependent), so the safest thing we can do is:

- If an in-memory format, insert an ordered column to use as an index
- Apply the full transform
- Order by the index column, then drop the index column

We don't do this when the input is already an ibis table. Since
scikit-learn will never pass in an ibis table and ibis tables don't have
an implicit order, we don't have an order to maintain. For users trying
to fit models using ibis directly they'll include `X` and `y` together
in the same table anyway (so a reordering of rows won't result in a
mismatch).