Plumb through `y` in `fit` and `fit_transform`
Previously all steps would only rely on `X` during fitting, so we
completely ignored `y` if it was passed in. We now have at least one
desired `Step` that relies on `y` during the fitting process, which
means we need to start properly handling `y` in `fit`/`fit_transform`.

This PR handles:
- normalizing `X` and `y` to a single `Table` from a variety of possible
  inputs
- Plumbing through `X` and `y` metadata through `fit` and
  `fit_transform`.

Still needs a bunch of tests, but I believe this works already in
general.